+++ shared/src/commonMain/kotlin/hr/dtakac/prognoza/shared/domain/GetAllForecastProviders.kt
@@ -1,0 +1,7 @@
+package hr.dtakac.prognoza.shared.domain
+
+import hr.dtakac.prognoza.shared.entity.ForecastProvider
+
+class GetAllForecastProviders {
+    suspend operator fun invoke(): List<ForecastProvider> = ForecastProvider.values().toList()
+}