+++ shared/src/commonMain/kotlin/hr/dtakac/prognoza/shared/domain/GetForecastProvider.kt
@@ -1,0 +1,10 @@
+package hr.dtakac.prognoza.shared.domain
+
+import hr.dtakac.prognoza.shared.domain.data.SettingsRepository
+import hr.dtakac.prognoza.shared.entity.ForecastProvider
+
+class GetForecastProvider internal constructor(
+    private val settingsRepository: SettingsRepository
+){
+    suspend operator fun invoke(): ForecastProvider = settingsRepository.getForecastProvider()
+}