+++ shared/src/commonMain/kotlin/hr/dtakac/prognoza/shared/entity/ForecastProvider.kt
@@ -1,0 +1,5 @@
+package hr.dtakac.prognoza.shared.entity
+
+enum class ForecastProvider {
+    MET_NORWAY, OPEN_METEO
+}